requirements.txt
geojson
numpy
onedrivesdk
opencv-python
openpyxl
pandas
python-dotenv
